Extract VariableDeclarationModification in its file
